Add more information to integrity test error message
When the integrity test encounters a file that does not contain
the expected data, it lists the data that it expected to find
in terms of writes to the file.

Now the test also displays a list of "raw" writes that includes
writes that have been truncated away, or completely overwritten
by other writes.

The test also now displays the pid because it is used as the
initial random seed.

Signed-off-by: Adrian Hunter <ext-adrian.hunter@nokia.com>