ubi-utils: get rid of compile warnings
Just silly hacks. Also remove udevsettle() invocation from UBI
library. If it must be called, it is not library's business
anyway. Wa added it to make scripts which use ubimkvol utility
run, so it is the only caller that really needs this. So
just move the cruft to the utility.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>