ubi-utils: introduce ubinfo utility
Add new handy UBI utility which prints various type of UBI information.

This commit also includes a lot of fixes and cleanups in libubi, and
other utilities. It was quite complex to separate this all out and
I figured that nobody anyway would really need this, and decided to
save my time for more useful things.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>