ubi-utils: add ubiupdate
Add an 'ubiupdate' utility which is basically a changed and cleaned-up
version of ubiupdatevol. It is also consistent with other ubi tools.

ubiupdatevol should go away, because it has features which should not
be present in tools for end users. If one needs debugging capabilities,
he should create a separate program.

So, I do not delete ubiupdatevol, but it will be moved somwthere later.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>