ubi-utils: remove all old tools
Remove all old tools because I cannot maintain them and the original
authors do not seem to have time for this. Some of the tools do not
work properly, some are just vague and undocumented and seem to be
oriented to the environment of the IBM guys. Nevertheless, I'll
return the tool as is in the next commit, becouse they are still
useful.

This commit also adds a ubinize utility to generate UBI images.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>