Makefile: Add Optional ZLIB and LZO CPPFLAGS and LDFLAGS
This adds support for allowing the specification of unique CPPFLAGS and
LDFLAGS for both LZO and ZLIB. In addition, it allows for independently
specifying both PREFIX and EXEC_PREFIX.

The former is useful in some cross-compilation environments where the
LZO and ZLIB versions being built against are not those installed in the
standard locations.

The latter is useful when installing mtd-utils in a multi-architecture
installation area.

Signed-off-by: Grant Erickson <gerickson@nuovations.com>
Signed-off-by: Josh Boyer <jwboyer@gmail.com>