ubiformat: reverse question about vid offset change
Users can pass the -O option to ubiformat to override the VID header
offset on the command-line. If for some reason, the new offset does
not match what is on flash, ubiformat currently prompts the user if
they'd rather use the *old* offsets.  This means that calling
ubiformat with --yes will not do what was requested on the
command-line.

This patch reverses the question so --yes is actually useful for
running in batch-mode.

Signed-off-by: Bernard Blackham <bernard@largestprime.net>
Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>