ubinize: fix static volumes generation
When static volumes is generated, the 'used_ebs' field of VID
header has to be calculated based on image file size, not on
the volume size. Fix this.

This patch also improves error reporting a little and re-arrages
the code so that all the validation is done in 'read_section()'

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>