lib: fix libcrc32 generation
libcrc32.a was generated without external symbols: nm -g libcrc32.a shoed
nothing. This patch fixes it to make libcrc32.a build to be the same as
libmtd.a. Frankly, I do not know why this happened and why this patch fixes
the issue, sorry for my ignorance. But this works.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>