mtd-utils: xalloc: simplify/unify error messages
I'm not sure that if we actually are out of memory that declaring the
failing allocation size is useful in the output.  So use the same
simple string in every error message to cut down on size (there will
only be one copy of this at runtime).  Size is a much more common
concern than handling OOM issues which most likely aren't the fault
of mtd-utils in the first place.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>