fs-tests: integck: improve remountability logic
The integck test assumes that if the FS is rootfs then it cannot
re-mount it, otherwise it can. However, this is not true because the
tested FS can be not remountable even if it is not rootfs, e.g., if
'integck' lives on this FS and is executed from it.

This patch improves the integck logic and actually checks wheter it
is possible to re-mount the FS before starting the test.

Also, it makes sure that power cut testing is run only on re-mountable
FS, because we have to be able to clean the "corrupted" state of the
file-system.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>