fs-tests: integck: implement MTD reattaching
This patch adds -m <mtdnum> option to integck and teaches integck to re-attach
the MTD device to UBI in case of an emulated power cut event. This is needed
for the new UBI power cut and unstable bits emulation infrastructure: when UBI
emulates a power cut the only way to recover form this is to re-attach the MTD
device.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>