fs-tests: integck: EIO can also be return in power-cut testing
Currently when we are doing UBIFS power cut emulation testing we assume that
EROFS is returned in case of an emulated power cut event. However, this is
not true and sometimes Linux returns EIO to user-space instead. Namely, this
happens when we are waiting for write-back to finish. Here is a call trace
form kernel 2.6.39-rc7:

Call Trace:
 [<ffffffff810aaa77>] filemap_write_and_wait_range+0x78/0xb1
 [<ffffffff8110f34d>] vfs_fsync_range+0x38/0xd3
 [<ffffffff8110f42b>] generic_write_sync+0x43/0x65
 [<ffffffff810ab0c9>] generic_file_aio_write+0xaa/0xf2
 [<ffffffffa02224c9>] ubifs_aio_write+0x13a/0x1a1 [ubifs]
 [<ffffffff810ed71f>] do_sync_write+0xcb/0x141
 [<ffffffff810ee016>] vfs_write+0xb0/0x150
 [<ffffffff810ee175>] sys_write+0x48/0x72
 [<ffffffff813e0692>] system_call_fastpath+0x16/0x1b

'filemap_fdatawait_range()' returns hard-coded -EIO if a page write-back
happens with whatever error code.

Thus, this patch teaches integck to handle EIO when doing power cut emulation
testing.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>