fs-tests: integck: print error message on error
Currently pcv macro print the error message only if we are not doing power cut
testing or if we have -v flag. But if we run without -v and an error happen
and the error code is not EROFS/EIO, pcv() does not print anything. This patch
makes it print the error message in that case as well.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>