flash_{lock,unlock}: merge functionality
The flash_lock util has a bit of extra argument checking, and it supports
a magic value of "-1" to mean "all blocks".

The flash_unlock util supports automatic 2nd/3rd arguments to unlock the
whole flash.  It also supports multiple bases (not just hex) for selecting
the range of the device to unlock.

So tweak both utilities so that they have equivalent functionality again
by adding the missing features to each.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>